Implement autocorrect based on ggml model

+++ java/src/org/futo/inputmethod/latin/GGMLDictionary.java
@@ -1,32 +1,17 @@
 package org.futo.inputmethod.latin;
 
-import static org.futo.inputmethod.latin.BinaryDictionary.DICTIONARY_MAX_WORD_LENGTH;
-
 import android.content.Context;
-import android.os.Build;
-import android.util.SparseArray;
-
-import androidx.annotation.RequiresApi;
 
 import org.futo.inputmethod.latin.common.ComposedData;
-import org.futo.inputmethod.latin.common.Constants;
-import org.futo.inputmethod.latin.common.FileUtils;
 import org.futo.inputmethod.latin.common.InputPointers;
 import org.futo.inputmethod.latin.settings.SettingsValuesForSuggestion;
-import org.futo.inputmethod.latin.utils.WordInputEventForPersonalization;
 
-import java.io.ByteArrayOutputStream;
 import java.io.File;
 import java.io.FileOutputStream;
-import java.io.FileWriter;
 import java.io.IOException;
 import java.io.InputStream;
 import java.io.OutputStream;
-import java.nio.file.Files;
-import java.nio.file.Paths;
-import java.security.MessageDigest;
 import java.util.ArrayList;
-import java.util.Arrays;
 import java.util.Locale;
 
 
@@ -34,20 +19,17 @@ import java.util.Locale;
 // Still kind of unsure. Maybe we should integrate more with BinaryDictionary
 // sort of like: P(word) = P(word) * P_TransformerLM( tokenize(word)[0] )
 
-// Step 1. Suggest next word based on the last three words in ngramContext
-// Step 2. Suggest next word based on the full previous sentence
-// Step 3. Suggest correction based on composeddata and proximityinfohandle
 public class GGMLDictionary extends Dictionary {
     long mNativeState = 0;
 
-    private String getPathToModelResource(Context context, int resource) {
+    private String getPathToModelResource(Context context, int resource, boolean forceDelete) {
         File outputDir = context.getCacheDir();
         File outputFile = new File(outputDir, "ggml-model-" + String.valueOf(resource) + ".bin");
 
-        if(outputFile.exists()) {
+        if(forceDelete && outputFile.exists()) {
             outputFile.delete();
         }
-        {
+        if((!outputFile.exists()) || forceDelete){
             // FIXME: We save this to a random temporary file so that we can have a path instead of an InputStream
             InputStream is = context.getResources().openRawResource(resource);
 
@@ -76,9 +58,17 @@ public class GGMLDictionary extends Dictionary {
     public GGMLDictionary(Context context, String dictType, Locale locale) {
         super(dictType, locale);
 
-        String modelPath = getPathToModelResource(context, R.raw.pythia_160m_q4_0);
-
+        String modelPath = getPathToModelResource(context, R.raw.pythia_160m_q4_0, false);
         mNativeState = openNative(modelPath, 0, 0, false);
+
+        if(mNativeState == 0){
+            modelPath = getPathToModelResource(context, R.raw.pythia_160m_q4_0, true);
+            mNativeState = openNative(modelPath, 0, 0, false);
+        }
+
+        if(mNativeState == 0){
+            throw new RuntimeException("Failed to load pythia_160m model");
+        }
     }
 
     @Override
@@ -101,15 +91,36 @@ public class GGMLDictionary extends Dictionary {
         inputSize = inputPointers.getPointerSize();
 
         String context = " " + ngramContext.extractPrevWordsContext().replace(NgramContext.BEGINNING_OF_SENTENCE_TAG, " ").trim();
+        if(!ngramContext.fullContext.isEmpty()) {
+            context = " " + ngramContext.fullContext.trim();
+        }
+
+        String partialWord = composedData.mTypedWord;
+
+        if(!partialWord.isEmpty() && context.endsWith(partialWord)) {
+            context = " " + context.substring(0, context.length() - partialWord.length()).trim();
+        }
+
+        if(!partialWord.isEmpty()) {
+            partialWord = " " + partialWord.trim();
+        }
+
         System.out.println("Context for ggml is " + context);
-        String[] outStrings = new String[256];
+        System.out.println("partialWord is " + partialWord);
+
+
+        int maxResults = 128;
+        int[] outProbabilities = new int[maxResults];
+        String[] outStrings = new String[maxResults];
 
         // TOOD: Pass multiple previous words information for n-gram.
-        getSuggestionsNative(mNativeState, proximityInfoHandle, context, outStrings);
+        getSuggestionsNative(mNativeState, proximityInfoHandle, context, partialWord, outStrings, outProbabilities);
 
         final ArrayList<SuggestedWords.SuggestedWordInfo> suggestions = new ArrayList<>();
-        for(int i=0; i<3; i++) {
-            suggestions.add(new SuggestedWords.SuggestedWordInfo( outStrings[i], context, 10, 1, this, 0, 0 ));
+        for(int i=0; i<maxResults; i++) {
+            if(outStrings[i] == null) continue;
+
+            suggestions.add(new SuggestedWords.SuggestedWordInfo( outStrings[i].trim(), context, outProbabilities[i], 1, this, 0, 0 ));
         }
         return suggestions;
     }
@@ -141,5 +152,5 @@ public class GGMLDictionary extends Dictionary {
     private static native long openNative(String sourceDir, long dictOffset, long dictSize,
                                           boolean isUpdatable);
     private static native void closeNative(long dict);
-    private static native void getSuggestionsNative(long dict, long proximityInfo, String context, String[] strings);
+    private static native void getSuggestionsNative(long dict, long proximityInfoHandle, String context, String partialWord, String[] outStrings, int[] outProbs);
 }
